style the carousel
